# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ETask API Documentation",
    "description": "基于 AST 与反射契约混合自适应生成的 OpenAPI 3.0 规范，支持导入 Apifox / Postman 进行全量接口联调。",
    "version": "1.0.0"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/manager/create": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "创建任务",
        "description": "权限标识: add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskReq"
              },
              "example": {
                "cron_expr": "string",
                "execution_notifications": [
                  "item1",
                  "item2"
                ],
                "grpc_config": "string",
                "http_config": "string",
                "max_execution_seconds": 0,
                "metadata": "string",
                "name": "示例名称",
                "param_override_rules": [
                  "id_001",
                  "id_002"
                ],
                "program": "string",
                "retry_config": "string",
                "runner_id": 1,
                "schedule_params": "string",
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/delete/{id}": {
      "delete": {
        "tags": [
          "任务管理"
        ],
        "summary": "删除任务",
        "description": "权限标识: delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/detail/{id}": {
      "get": {
        "tags": [
          "任务管理"
        ],
        "summary": "任务详情",
        "description": "权限标识: get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/executions": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "执行记录",
        "description": "权限标识: executions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListExecutionsReq"
              },
              "example": {
                "limit": 10,
                "offset": 0,
                "task_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/executions/{id}/parameters": {
      "get": {
        "tags": [
          "任务管理"
        ],
        "summary": "执行参数",
        "description": "权限标识: execution_parameters",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/executions/{id}/terminate": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "终止执行",
        "description": "权限标识: terminate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminateExecutionReq"
              },
              "example": {
                "reason": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/list": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "任务列表",
        "description": "权限标识: view",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageReq"
              },
              "example": {
                "limit": 10,
                "offset": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/logs": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "任务日志",
        "description": "权限标识: logs",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetLogsReq"
              },
              "example": {
                "execution_id": 1,
                "limit": 10,
                "min_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/run": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "运行任务",
        "description": "权限标识: start",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunTaskReq"
              },
              "example": {
                "cron_expr": "string",
                "id": 1,
                "param_overrides": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/stop/{id}": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "停止任务",
        "description": "权限标识: stop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/manager/update": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "更新任务",
        "description": "权限标识: edit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskReq"
              },
              "example": {
                "cron_expr": "string",
                "execution_notifications": [
                  "item1",
                  "item2"
                ],
                "grpc_config": "string",
                "http_config": "string",
                "id": 1,
                "max_execution_seconds": 0,
                "metadata": "string",
                "name": "示例名称",
                "param_override_rules": [
                  "id_001",
                  "id_002"
                ],
                "program": "string",
                "retry_config": "string",
                "runner_id": 1,
                "schedule_params": "string",
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
    "/api/streams/manager/executions/{id}/logs": {
      "get": {
        "tags": [
          "任务管理"
        ],
        "summary": "查看执行日志流",
        "description": "权限标识: execution_logs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
    "/api/streams/manager/tasks/{id}/executions": {
      "get": {
        "tags": [
          "任务管理"
        ],
        "summary": "订阅任务执行事件",
        "description": "权限标识: execution_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "CreateTaskReq": {
        "type": "object",
        "properties": {
          "cron_expr": {
            "type": "string",
            "description": "cron 表达式（定时任务必填，一次性任务可选用于定时触发）",
            "example": "string"
          },
          "execution_notifications": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "item1",
              "item2"
            ]
          },
          "grpc_config": {
            "type": "string",
            "example": "string"
          },
          "http_config": {
            "type": "string",
            "example": "string"
          },
          "max_execution_seconds": {
            "type": "integer",
            "description": "最大执行秒数，默认24小时",
            "example": 0
          },
          "metadata": {
            "type": "string",
            "description": "任务参数元数据",
            "example": "string"
          },
          "name": {
            "type": "string",
            "example": "示例名称"
          },
          "param_override_rules": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "id_001",
              "id_002"
            ]
          },
          "program": {
            "type": "string",
            "description": "程序来源",
            "example": "string"
          },
          "retry_config": {
            "type": "string",
            "example": "string"
          },
          "runner_id": {
            "type": "integer",
            "example": 1
          },
          "schedule_params": {
            "type": "string",
            "description": "调度参数（如分页偏移量、处理进度等）",
            "example": "string"
          },
          "type": {
            "type": "string",
            "description": "任务类型: RECURRING-定时任务, ONE_TIME-一次性任务",
            "example": "string"
          }
        },
        "example": {
          "cron_expr": "string",
          "execution_notifications": [
            "item1",
            "item2"
          ],
          "grpc_config": "string",
          "http_config": "string",
          "max_execution_seconds": 0,
          "metadata": "string",
          "name": "示例名称",
          "param_override_rules": [
            "id_001",
            "id_002"
          ],
          "program": "string",
          "retry_config": "string",
          "runner_id": 1,
          "schedule_params": "string",
          "type": "string"
        }
      },
...
      "GetLogsReq": {
        "type": "object",
        "properties": {
          "execution_id": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "min_id": {
            "type": "integer",
            "example": 1
          }
        },
        "example": {
          "execution_id": 1,
          "limit": 10,
          "min_id": 1
        }
      },
...
      "ListExecutionsReq": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer",
            "example": 10
          },
          "offset": {
            "type": "integer",
            "example": 0
          },
          "task_id": {
            "type": "integer",
            "example": 1
          }
        },
        "example": {
          "limit": 10,
          "offset": 0,
          "task_id": 1
        }
      },
...
      "PageReq": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer",
            "example": 10
          },
          "offset": {
            "type": "integer",
            "example": 0
          }
        },
        "example": {
          "limit": 10,
          "offset": 0
        }
      },
...
      "RunTaskReq": {
        "type": "object",
        "properties": {
          "cron_expr": {
            "type": "string",
            "example": "string"
          },
          "id": {
            "type": "integer",
            "example": 1
          },
          "param_overrides": {
            "type": "string",
            "example": "string"
          }
        },
        "example": {
          "cron_expr": "string",
          "id": 1,
          "param_overrides": "string"
        }
      },
...
      "TerminateExecutionReq": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "string"
          }
        },
        "example": {
          "reason": "string"
        }
      },
...
      "UpdateTaskReq": {
        "type": "object",
        "properties": {
          "cron_expr": {
            "type": "string",
            "description": "cron 表达式（定时任务必填，一次性任务可选用于定时触发）",
            "example": "string"
          },
          "execution_notifications": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "item1",
              "item2"
            ]
          },
          "grpc_config": {
            "type": "string",
            "example": "string"
          },
          "http_config": {
            "type": "string",
            "example": "string"
          },
          "id": {
            "type": "integer",
            "example": 1
          },
          "max_execution_seconds": {
            "type": "integer",
            "description": "最大执行秒数，默认24小时",
            "example": 0
          },
          "metadata": {
            "type": "string",
            "description": "任务参数元数据",
            "example": "string"
          },
          "name": {
            "type": "string",
            "example": "示例名称"
          },
          "param_override_rules": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "id_001",
              "id_002"
            ]
          },
          "program": {
            "type": "string",
            "description": "程序来源",
            "example": "string"
          },
          "retry_config": {
            "type": "string",
            "example": "string"
          },
          "runner_id": {
            "type": "integer",
            "example": 1
          },
          "schedule_params": {
            "type": "string",
            "description": "调度参数",
            "example": "string"
          },
          "type": {
            "type": "string",
            "description": "任务类型: RECURRING-定时任务, ONE_TIME-一次性任务",
            "example": "string"
          }
        },
        "example": {
          "cron_expr": "string",
          "execution_notifications": [
            "item1",
            "item2"
          ],
          "grpc_config": "string",
          "http_config": "string",
          "id": 1,
          "max_execution_seconds": 0,
          "metadata": "string",
          "name": "示例名称",
          "param_override_rules": [
            "id_001",
            "id_002"
          ],
          "program": "string",
          "retry_config": "string",
          "runner_id": 1,
          "schedule_params": "string",
          "type": "string"
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "请输入登录后获取的 JWT Bearer Token 进行接口调用鉴权"
      }
    }
  }
}